cloud_functions/document_trigger/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
cloudevents==1.*
httpx==0.26.*
